pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tether-engine"
version = "0.8.0"
description = "An impact-analysis and edge-case engine for coding agents"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "watchdog>=4.0.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
]

[project.scripts]
tether = "tether.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
ollama = [
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tether"]

[tool.hatch.build.targets.sdist]
include = ["src/tether"]

[tool.pytest.ini_options]
testpaths = ["tests"]
